moves config, edits image overlays
