+++ src/config/config.js
@@ -4,5 +4,50 @@ module.exports = {
 		large: 1200,
 		tablet: 768,
 		phone: 480
-	}
+	},
+
+	pages: [
+		'portfolio', 
+		'about',
+		'blog', 
+		'contact'
+	],
+
+	socialMedia: [
+		{
+			url: 'http://codepen.io/ryanfiller89/',
+			name: 'Codepen',
+			color: '#fcd000'
+		},
+		{
+			url: 'https://github.com/ryanfiller',
+			name: 'Github',
+			color: '#6cc644'
+		},
+		{
+			url: 'https://www.linkedin.com/in/ryanfiller',
+			name: 'LinkedIn',
+			color: '#0077b5'
+		},
+		{
+			url: 'https://www.facebook.com/ryan.filler',
+			name: 'Facebook',
+			color: '#3b5998'
+		},
+		{
+			url: 'https://twitter.com/ryanfiller_',
+			name: 'Twitter',
+			color: '#1da1f2'
+		},
+		{
+			url: 'https://www.instagram.com/ryanfiller_/',
+			name: 'Instagram',
+			color: '#bc2a8d'
+		},
+		{
+			url: 'https://dribbble.com/ryanoh',
+			name: 'Dribbble',
+			color: '#ea4c89'
+		},
+	],
 };
