+++ src/styles/config/_colors.scss
@@ -16,24 +16,29 @@ $shadow-inset:
     inset 0px 2px 10px rgba(255, 0, 0, 0.30),
     inset 2px 0px 10px rgba(255, 0, 0, 0.30);
 
+$pixel: .25rem;
+
+$r: rgba(255,0,0,.15);
+$g: rgba(0,255,0,.15);
+$b: rgba(0,0,255,.15);
+$rgb: 
+repeating-linear-gradient(
+    to bottom,
+    $r 0*$pixel, $r 1*$pixel,
+    $g 1*$pixel, $g 2*$pixel,
+    $b 2*$pixel, $b 3*$pixel
+);
 
 @mixin pixels {
     background-image: url('/images/site-assets/screendoor.png');
-    background-position: center center;
+    background-size: $pixel;
 }
 
-$rgb: repeating-linear-gradient(
-    to right,
-    rgba(0,0,0,.5), rgba(0,0,0,.5) .125rem,
-    rgba(255,255,255,.5) .25rem, rgba(255,255,255,.25) .5rem
-),
-repeating-linear-gradient(
-    to bottom,
-    rgba(255,0,0,.5) 0rem, rgba(255,0,0,.5) .25rem,
-    rgba(0,255,0,.5) .25rem, rgba(0,255,0,.5) .5rem,
-    rgba(0,0,255,.5) .5rem, rgba(0,0,255,.5) .75rem,
-    lime .75rem
-);
+@mixin rgb-pixels {
+    background-image: url('/images/site-assets/screendoor.png'), $rgb;
+    background-size: $pixel;
+    mix-blend-mode: hard-light;
+}
 
 @mixin image-overlay {
     position: relative;
