+++ src/styles/pages/_blog-item.scss
@@ -19,7 +19,7 @@ $content-offset: 10rem;
                 right: 0;
                 bottom: 0;
                 left: 0;
-                @include pixels;
+                @include rgb-pixels;
                 opacity: .5;
             }
